pyproject.toml
[tool.poetry]
name = "hyperliquid-monitor"
version = "0.1.2"
description = "A Hyperliquid trade monitor package"
authors = ["Yakitori"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/oni-giri/hyperliquid-monitor"
repository = "https://github.com/oni-giri/hyperliquid-monitor"
keywords = ["hyperliquid", "trading", "cryptocurrency", "monitor", "dex"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business :: Financial :: Investment",
]
packages = [{include = "hyperliquid_monitor", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
hyperliquid-python-sdk = "^0.8.0"
python-dotenv = "^1.0.0"
discord = "^2.3.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-mock = "^3.12.0"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
freezegun = "^1.2.2"

[tool.poetry.scripts]
hyperliquid-discord = "hyperliquid_monitor.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=hyperliquid_monitor --cov-report=term-missing"